requirements.txt
playwright==1.44.0
pandas==2.2.2
openpyxl==3.1.4
python-dotenv==1.0.1
httpx==0.27.0
schedule==1.2.2
anthropic==0.25.0